refactor: apply ktlint formatting to presentation layer
