+++ presentation/src/main/kotlin/com/dobby/api/controller/MemberController.kt
@@ -1,6 +1,10 @@
 package com.dobby.api.controller
 
-import com.dobby.api.dto.request.member.*
+import com.dobby.api.dto.request.member.ParticipantSignupRequest
+import com.dobby.api.dto.request.member.ResearcherSignupRequest
+import com.dobby.api.dto.request.member.UpdateResearcherInfoRequest
+import com.dobby.api.dto.request.member.UpdateParticipantInfoRequest
+import com.dobby.api.dto.request.member.DeleteMemberRequest
 import com.dobby.api.dto.response.member.DefaultResponse
 import com.dobby.api.dto.response.member.ParticipantInfoResponse
 import com.dobby.api.dto.response.member.ResearcherInfoResponse
